Moved a few utility functions from scons process' plugin_util.py to common_util.py. This will allow using them also from the Apio process.
