+++ .vscode/settings.json
@@ -125,6 +125,7 @@
         "spammy",
         "srcs",
         "Sucess",
+        "synthesizable",
         "testbench",
         "testbenches",
         "testpypi",
