+++ test/unit_tests/scons/testing.py
@@ -33,8 +33,6 @@ environment {
   debug_level: 1
   yosys_path: "/Users/user/.apio/packages/oss-cad-suite/share/yosys"
   trellis_path: "/Users/user/.apio/packages/oss-cad-suite/share/trellis"
-  build_all_path: "_build"
-  build_env_path: "_build/default"
 }
 apio_env_params {
   env_name: "default"
